Change library name to seriate. Simplify how library is required by consuming code.
