+++ README.md
@@ -1,4 +1,4 @@
-#ms-squeal
+#Seriate
 =========
 
 A cross platform node module for Microsoft SQL Server based on [node-mssql](https://www.npmjs.org/package/mssql)
@@ -51,7 +51,7 @@ This method returns a `TransactionContext` instance, and allows you to add 1 or 
 
 Here's an example of using a plain context to read a table:
 
-	squeal.getPlainContext( {
+	seriate.getPlainContext( {
 		user: 'username',
 		password: 'pwd',
 		server: '127.0.0.1',
@@ -95,7 +95,7 @@ The `end` method of a `SqlContext` instance takes a callback which receives a `s
 
 Here's an example of using a plain context to read a table, and then use data from that read to determine details about the next step:
 
-	squeal.getPlainContext( {
+	seriate.getPlainContext( {
 		user: 'username',
 		password: 'pwd',
 		server: '127.0.0.1',
@@ -132,7 +132,7 @@ The above example shows both `step` approaches side-by-side.
 ###getTransactionContext(connectionConfig)
 The `getTransactionContext` method returns a `TransactionContext` instance - which for the most part is nearly identical to a `SqlContext` instance - however, a transaction is started as the context begins its work, and you have the option to commit or rollback in the `end` method's callback. For example:
 
-	squeal.getTransactionContext( {
+	seriate.getTransactionContext( {
 		user: 'username',
 		password: 'pwd',
 		server: '127.0.0.1',
@@ -181,7 +181,7 @@ You can see that the main difference between a `SqlContext` and `TransactionCont
 ###executeTransaction(connectionConfig, queryOptions)
 This is a shortcut method to getting a `TransactionContext` instance to execute one step. It returns a promise, and the `result` argument that's normally fed to the `end` method's callback is passed to the success handler of the promise, and any errors are passed to the error handler. For example:
 
-	squeal.executeTransaction( config, {
+	seriate.executeTransaction( config, {
 		procedure: "UpdateCustomer",
 		params: {
 			customerid: {
@@ -210,7 +210,7 @@ This is a shortcut method to getting a `TransactionContext` instance to execute 
 ###execute(connectionConfig, queryOptions)
 This is a shortcut method to getting a `SqlContext` instance to execute one step. It returns a promise, and the `result` argument that's normally fed to the `end` method's callback is passed to the success handler of the promise, and any errors are passed to the error handler. For example:
 
-	squeal.execute( config, {
+	seriate.execute( config, {
 		preparedSql: "select * from someTable where id = @id",
 		params: {
 			id: {
