+++ src/index.js
@@ -1,42 +1,8 @@
-var when = require( 'when' );
-var sql;
-var TransactionContext;
-var SqlContext;
+var sql = require( 'mssql' );
+var Monologue = require( 'monologue.js' );
+var machina = require( 'machina' )();
+var SqlContext = require( '../src/sqlContext.js' )( sql, Monologue, machina );
+var TransactionContext = require( '../src/transactionContext.js' )( sql, SqlContext );
+var seriate = require( '../src/main.js' )( SqlContext, TransactionContext );
 
-function promisify( context, queryOptions ) {
-	context.step( 'result', queryOptions );
-	return when.promise( function( resolve, reject, notify ) {
-		context
-			.end( resolve )
-			.error( reject )
-			.on( 'data', notify );
-	} );
-}
-
-module.exports = function( mssql, SqlContextCtor, TransactionContextCtor ) {
-	sql = mssql;
-	SqlContext = SqlContextCtor;
-	TransactionContext = TransactionContextCtor;
-	return {
-		getTransactionContext: function( config ) {
-			return new TransactionContext( {
-				connectionCfg: config
-			} );
-		},
-		getPlainContext: function( config ) {
-			return new SqlContext( {
-				connectionCfg: config
-			} );
-		},
-		executeTransaction: function( connCfg, queryOptions ) {
-			return promisify( new TransactionContext( {
-				connectionCfg: connCfg
-			} ), queryOptions );
-		},
-		execute: function( connCfg, queryOptions ) {
-			return promisify( new SqlContext( {
-				connectionCfg: connCfg
-			} ), queryOptions );
-		}
-	};
-};
+module.exports = seriate;
